Move Station stuff to its own directory
So we can later on package this
