+++ PyQT5_fbs/src/main/python/station_tools/filt_wts.txt
