+++ PyQT5_fbs/src/main/python/station_tools/filter_weights.py
